image_vault: Remove instance directory only if cd succeeds
Fixes #254

+++ src/daemon/default_vm_image_vault.cpp
@@ -391,8 +391,8 @@ void mp::DefaultVMImageVault::remove(const std::string& name)
         return;
 
     QDir instance_dir{instances_dir};
-    instance_dir.cd(QString::fromStdString(name));
-    instance_dir.removeRecursively();
+    if (instance_dir.cd(QString::fromStdString(name)))
+        instance_dir.removeRecursively();
 
     instance_image_records.erase(name);
     persist_instance_records();
